refactor: separated ecents, tickets and wishlist
