+++ events/custom_filters.py
@@ -1,6 +1,6 @@
 import django_filters
 
-from Event_Management.events_tickets.models import Event
+from events.models import Event
 
 
 class EventFilter(django_filters.FilterSet):
