+++ events/models.py
@@ -1,15 +1,7 @@
 from django.db import models
 
 from Event_Management.settings import AUTH_USER_MODEL
-
-
-class TicketType(models.Model):
-    name = models.CharField(max_length=20, null=False, blank=False)
-    is_active = models.BooleanField(default=True, verbose_name="active")
-
-    def delete(self, *args, **kwargs):
-        self.is_active = False
-        self.save()
+from tickets.models import TicketType
 
 
 class Event(models.Model):
@@ -49,23 +41,3 @@ class EventTicketType(models.Model):
 class Photo(models.Model):
     event = models.ForeignKey(Event, related_name="event_img", on_delete=models.CASCADE)
     image = models.URLField("url", null=False)
-
-
-class Ticket(models.Model):
-    event = models.ForeignKey(Event, on_delete=models.CASCADE)
-    customer = models.ForeignKey(AUTH_USER_MODEL, on_delete=models.CASCADE)
-    ticket_type = models.ForeignKey(
-        TicketType, verbose_name="ticket", on_delete=models.CASCADE
-    )
-    price = models.DecimalField(max_digits=7, decimal_places=2)
-    is_active = models.BooleanField(default=True, verbose_name="active")
-    archive = models.BooleanField(default=False, verbose_name="archive")
-
-    def delete(self, *args, **kwargs):
-        self.is_active = False
-        self.save()
-
-
-class Wishlist(models.Model):
-    created_by = models.ForeignKey(AUTH_USER_MODEL, on_delete=models.CASCADE)
-    event = models.ForeignKey(Event, on_delete=models.CASCADE)
