+++ tickets/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class EventsTicketsConfig(AppConfig):
+class TicketsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "events_tickets"
+    name = "tickets"
