+++ tickets/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path
-
-from Event_Management.events_tickets import views
+from tickets import views
 
 urlpatterns = [
     path("", views.TicketLC.as_view(), name="LC-ticket"),
